apps/discord-bot/pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "google-genai>=1.26.0",
    "python-dotenv>=1.1.1",
]
